URB-3020. Add translation and add contextual title to the form from send email action
