+++ news/URB-3020.feature
@@ -1,0 +1,2 @@
+Add translation and add contextual title to the form from send email action
+[jchandelle]
