+++ src/Products/urban/locales/urban.pot
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
-"POT-Creation-Date: 2024-09-10 13:36+0000\n"
+"POT-Creation-Date: 2024-10-03 08:07+0000\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI +ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -18,6 +18,10 @@ msgstr ""
 #: ../browser/offdays_settings.py:72
 #: ../browser/schedule_settings.py:17
 msgid ""
+msgstr ""
+
+#: ../send_mail_action/forms.py:42
+msgid " and ${last_rule}"
 msgstr ""
 
 #: ../content/licence/GenericLicence.py:1840
@@ -615,6 +619,10 @@ msgstr ""
 msgid "Internal opinion services"
 msgstr ""
 
+#: ../send_mail_action/forms.py:19
+msgid "Licence Files"
+msgstr ""
+
 #: ../browser/urbaneventviews.py:285
 msgid "Listing file (claimants)"
 msgstr ""
@@ -760,6 +768,10 @@ msgstr ""
 
 #: ../browser/templates/searchparcels.pt:34
 msgid "Search parcel(s)"
+msgstr ""
+
+#: ../send_mail_action/forms.py:20
+msgid "Select all files from this event or the parent licence you whant to send"
 msgstr ""
 
 #: ../docgen/UrbanTemplate.py:29
